harden api_key redaction and validate client tunables
Scrub the actual key value (raw and url-encoded) from non-2xx error
bodies before they hit the envelope, so an upstream that echoes the
credential in any shape (JSON, colon, quoted), not just the api_key=
query form the pattern catches, cannot leak it.

Widen the response cap to a long with a 256 MiB ceiling and fail fast on
a non-positive/oversized cap or a null/negative rate-limit budget,
instead of silently coercing some knobs while the cache bounds already
crash startup.

Tests cover the JSON/encoded redaction, the exact-cap and cap+1
boundary, and rejection of bad tunables.

+++ src/main/java/com/raorbit/opendota/config/OpenDotaProperties.java
@@ -27,7 +27,7 @@ public class OpenDotaProperties {
      * exceeds this is aborted rather than buffered, so a hostile or misbehaving
      * upstream cannot exhaust the heap.
      */
-    private int maxResponseBytes = 16 * 1024 * 1024;
+    private long maxResponseBytes = 16L * 1024 * 1024;
 
     public int getCacheMaxEntries() {
         return cacheMaxEntries;
@@ -53,11 +53,11 @@ public class OpenDotaProperties {
         this.rateLimitBudget = rateLimitBudget;
     }
 
-    public int getMaxResponseBytes() {
+    public long getMaxResponseBytes() {
         return maxResponseBytes;
     }
 
-    public void setMaxResponseBytes(int maxResponseBytes) {
+    public void setMaxResponseBytes(long maxResponseBytes) {
         this.maxResponseBytes = maxResponseBytes;
     }
 }
